Moved all commands over to cogs.  There is a Class in each module that organizes the commands.  Also cleaned up everything and removed any unneeded code.
