+++ bot/admin/strings.py
